Ignore errors from `npm ls` and `pip list`
`npm ls` fails if some peerDependencies are not installed
Since `showWebDependencies()` function is for debug purposes only, we catch possible promise reject of `npm()` and convert it into Promise.resolve() that it always succeeds. We do something similar in the CircleCI config with the `true` in `npm list --depth=1 || true`.

Since `showPythonTestDependencies()` is for debug purposes only, we catch possible errors and resolve it with status code `0`.